+++ src/OpenGL/libGLESv2/libGLESv3.cpp
@@ -2648,7 +2648,7 @@ GL_APICALL void GL_APIENTRY glCopyBufferSubData(GLenum readTarget, GLenum writeT
 		{
 			return error(GL_INVALID_ENUM);
 		}
-		if((readOffset + size) > readBuffer->size())
+		if(static_cast<size_t>(readOffset + size) > readBuffer->size())
 		{
 			return error(GL_INVALID_VALUE);
 		}
@@ -2658,7 +2658,7 @@ GL_APICALL void GL_APIENTRY glCopyBufferSubData(GLenum readTarget, GLenum writeT
 		{
 			return error(GL_INVALID_ENUM);
 		}
-		if((writeOffset + size) > writeBuffer->size())
+		if(static_cast<size_t>(writeOffset + size) > writeBuffer->size())
 		{
 			return error(GL_INVALID_VALUE);
 		}
